+++ CHANGELOG.md
@@ -6,7 +6,7 @@ If you like this project and find it useful, please consider giving it a **star*
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
-## [0.1.0] - Dev branch
+## [0.1.0] - 2026-04-09
 
 ### Added